pyproject.toml
[tool.poetry]
name = "drf-apis-in-a-nutshell-ljohns635"
version = "0.1.0"
description = ""
authors = ["Latisha Johnson <latishaj966@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.7"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
